+++ package.json
@@ -53,7 +53,7 @@
     "moduleNameMapper": {
       "^_Demo(.*)$": "<rootDir>/src/_Demo$1",
       "^validations(.*)$": "<rootDir>/src/validations/src$1",
-      "^core(.*)$": "<rootDir>/src/core$1"
+      "^core(.*)$": "<rootDir>/src//core/Mobx-MVVM/src$1"
     }
   }
 }
