+++ src/_Demo/tabs/SimpleFieldsTab/store.js
@@ -1,6 +1,6 @@
 import PersonalInformation from './containers/PersonalInformation/store';
-import ComplexType from '../../../core/ComplexType';
-import modelProp from '../../../core/modelProp';
+import ComplexType from 'core/ComplexType';
+import modelProp from 'core/modelProp';
 
 class SimpleFieldsTab extends ComplexType {
   constructor() {
