+++ src/core/Mobx-MVVM/_tests_/complexType.spec.js
@@ -1,29 +1,32 @@
-import ValidationsManager from '../../validations/src/core/validationsManager';
-import PropertiesManager from 'core/PropertiesManager';
-import formObservableGenerator from '../../core/formObservableGenerator';
-import { modelPropGenerator } from '../../core/modelProp';
-jest.mock('../../core/modelProp', () => ({
-  default: require.requireActual('../../core/modelProp').default,
+import ValidationsManager from '../../../validations/src/core/validationsManager';
+import FormObservablesManager from '../src/propertiesManager/FormObservablesManager';
+import ModelPropsManager from '../src/propertiesManager/ModelPropsManager';
+
+import { modelPropGenerator } from '../src/modelProp';
+import formObservableGenerator from '../src/formObservableGenerator';
+
+jest.mock('../src/modelProp', () => ({
+  default: require.requireActual('../src/modelProp').default,
   modelPropGenerator: jest.fn()
 }));
-jest.mock('../../core/formObservableGenerator');
-jest.mock('../../validations/src/core/validationsManager');
-
-import ComplexTab from '../mocks/ComplexTab';
+jest.mock('../../../validations/src/core/validationsManager');
+jest.mock('../src/formObservableGenerator');
+import ComplexTab from './mocksExamples/ComplexTab';
 
 let customTab;
 const settings = { validations: [() => true] };
 beforeEach(() => {
   ValidationsManager.mockClear();
-  formObservableGenerator.mockClear();
   modelPropGenerator.mockClear();
+  formObservableGenerator.mockClear();
+  Object.defineProperty = jest.fn();
   customTab = new ComplexTab();
-  customTab.propertiesManager.validate = jest.fn(() => 
-     {
+  customTab.formObservablesManager.validate = jest.fn(() => {
+    return {
       isValid: true,
       message: ''
-    }
-  );
+    };
+  });
   customTab.validationsManager.validate = jest.fn(() => {
     return {
       isValid: true,
@@ -47,32 +50,25 @@ describe('ComplexType constructor', () => {
         );
       });
     });
-    test('propertiesManager', () => {
-      expect(customTab.propertiesManager).toBeInstanceOf(PropertiesManager);
+    test('FormObservablesManager', () => {
+      expect(customTab.formObservablesManager).toBeInstanceOf(
+        FormObservablesManager
+      );
+    });
+    test('ModelPropsManager', () => {
+      expect(customTab.modelPropsManager).toBeInstanceOf(ModelPropsManager);
     });
   });
 
-  describe('setPropertiesReferences ', () => {
-    beforeEach(() => {
-      customTab.propertiesManager.getPropertiesDescriptors = jest
-        .fn()
-        .mockReturnValue({ firstName: { value: 1 }, agreement: { value: 2 } });
-    });
-    test('call propertiesManager.getPropertiesDescriptors', () => {
-      customTab.setPropertiesReferences();
-      expect(
-        customTab.propertiesManager.getPropertiesDescriptors.mock.calls.length
-      ).toBe(1);
-    });
-    test('set the value from getPropertiesDescriptors in the complex by key', () => {
-      expect(ComplexTab.prototype.firstName).toEqual({});
-      customTab.setPropertiesReferences();
-
-      expect(customTab.firstName).toEqual({ value: 1 });
-    });
-  });
+  // describe('setPropertiesReferences ', () => {
+  //   test('set the value from getPropertiesDescriptors in the complex by key', () => {
+  //     customTab.firstName = 1;
+  //     customTab.setPropertiesReferences();
+  //     expect(customTab.firstName).toBe('yael');
+  //   });
+  // });
   describe('loop over _propertiesSettings', () => {
-    describe('call generateModelProp for modelProps properties', () => {
+    describe('call formObservableGenerator for formObservables properties', () => {
       test('agreement', () => {
         expect(formObservableGenerator.mock.calls[0][0]).toBeDefined();
         expect(formObservableGenerator.mock.calls[0][0].name).toBe(
@@ -87,9 +83,9 @@ describe('ComplexType constructor', () => {
         expect(
           formObservableGenerator.mock.calls[0][0].validationsManager
         ).toBe(customTab._propertiesSettings.agreement.validationsManager);
-        expect(formObservableGenerator.mock.calls[0][0].propertiesManager).toBe(
-          customTab.propertiesManager
-        );
+        expect(
+          formObservableGenerator.mock.calls[0][0].formObservablesManager
+        ).toBe(customTab.formObservablesManager);
       });
       test('firstName', () => {
         expect(formObservableGenerator.mock.calls[1][0]).toBeDefined();
@@ -105,9 +101,9 @@ describe('ComplexType constructor', () => {
         expect(
           formObservableGenerator.mock.calls[1][0].validationsManager
         ).toBe(customTab._propertiesSettings.firstName.validationsManager);
-        expect(formObservableGenerator.mock.calls[1][0].propertiesManager).toBe(
-          customTab.propertiesManager
-        );
+        expect(
+          formObservableGenerator.mock.calls[1][0].formObservablesManager
+        ).toBe(customTab.formObservablesManager);
       });
     });
     describe('call modelPropGenerator for modelProps properties', () => {
@@ -116,7 +112,7 @@ describe('ComplexType constructor', () => {
         expect(modelPropGenerator.mock.calls[0][0].name).toBe('agreement');
         expect(modelPropGenerator.mock.calls[0][0].descriptor).toBeDefined();
         expect(
-          modelPropGenerator.mock.calls[0][0].propertiesManager
+          modelPropGenerator.mock.calls[0][0].modelPropsManager
         ).toBeDefined();
       });
       test('firstName', () => {
@@ -124,7 +120,7 @@ describe('ComplexType constructor', () => {
         expect(modelPropGenerator.mock.calls[1][0].name).toBe('firstName');
         expect(modelPropGenerator.mock.calls[1][0].descriptor).toBeDefined();
         expect(
-          modelPropGenerator.mock.calls[1][0].propertiesManager
+          modelPropGenerator.mock.calls[1][0].modelPropsManager
         ).toBeDefined();
       });
       test('complex', () => {
@@ -132,7 +128,7 @@ describe('ComplexType constructor', () => {
         expect(modelPropGenerator.mock.calls[2][0].name).toBe('complex');
         expect(modelPropGenerator.mock.calls[2][0].descriptor).toBeDefined();
         expect(
-          modelPropGenerator.mock.calls[2][0].propertiesManager
+          modelPropGenerator.mock.calls[2][0].modelPropsManager
         ).toBeDefined();
       });
     });
@@ -142,24 +138,24 @@ describe('ComplexType constructor', () => {
 describe('initializeComplexProperties', () => {
   beforeEach(() => {
     customTab = new ComplexTab();
-    customTab.propertiesManager.setComplexProperty = jest.fn();
+    customTab.modelPropsManager.setComplexProperty = jest.fn();
     customTab.initializeComplexProperties();
   });
 
   describe('call setComplexProperty for every complex property', () => {
     test('only with all complex properties', () => {
       expect(
-        customTab.propertiesManager.setComplexProperty.mock.calls.length
+        customTab.modelPropsManager.setComplexProperty.mock.calls.length
       ).toBe(1);
     });
     test('params: propty name', () => {
       expect(
-        customTab.propertiesManager.setComplexProperty.mock.calls[0][0]
+        customTab.modelPropsManager.setComplexProperty.mock.calls[0][0]
       ).toBe('complex');
     });
     test('params: object with propety validate function', () => {
       expect(
-        customTab.propertiesManager.setComplexProperty.mock.calls[0][1]
+        customTab.modelPropsManager.setComplexProperty.mock.calls[0][1]
       ).toEqual({ validate: customTab.complex.validate });
     });
   });
@@ -177,7 +173,7 @@ describe('validate - validate itself and all childs', () => {
       };
       beforeEach(() => {
         customTab.validationsManager.validate = jest.fn(() => result);
-        customTab.propertiesManager.validate = jest.fn(() => result);
+        customTab.formObservablesManager.validate = jest.fn(() => result);
       });
       test('call validationState.setValidationState with result', () => {
         customTab.validate();
