+++ src/core/Mobx-MVVM/_tests_/typeVerifications.spec.js
@@ -1,4 +1,4 @@
-import assertParametersType from '../../core/typeVerifications';
+import assertParametersType from 'core/typeVerifications';
 import { assertPropTypes } from 'check-prop-types';
 import PropTypes from 'prop-types';
 jest.mock('check-prop-types');
