+++ src/core/Mobx-MVVM/src/mobxConfiguration.js
