+++ src/core/Mobx-MVVM/src/propertiesManager/ModelPropBehavior.js
