+++ src/core/Mobx-MVVM/src/propertiesManager/ModelPropsManager.js
